pyproject.toml
[project]
name = "2025-agronomo-virtual-chatbot-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.18",
    "openai>=1.61.1",
    "pymupdf>=1.25.3",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
    "langchain-community>=0.3.17"
]
